+++ ethsigner/core/src/main/java/tech/pegasys/ethsigner/core/Runner.java
@@ -24,6 +24,7 @@ import tech.pegasys.ethsigner.core.requesthandler.VertxRequestTransmitter;
 import tech.pegasys.ethsigner.core.requesthandler.VertxRequestTransmitterFactory;
 import tech.pegasys.ethsigner.core.requesthandler.internalresponse.EthAccountsResultProvider;
 import tech.pegasys.ethsigner.core.requesthandler.internalresponse.EthSignResultProvider;
+import tech.pegasys.ethsigner.core.requesthandler.internalresponse.EthSignTransactionResultProvider;
 import tech.pegasys.ethsigner.core.requesthandler.internalresponse.InternalResponseHandler;
 import tech.pegasys.ethsigner.core.requesthandler.passthrough.PassThroughHandler;
 import tech.pegasys.ethsigner.core.requesthandler.sendtransaction.DownstreamPathCalculator;
@@ -173,7 +174,11 @@ public class Runner {
     requestMapper.addHandler(
         "eth_sign",
         new InternalResponseHandler<>(responseFactory, new EthSignResultProvider(signerProvider)));
-
+    requestMapper.addHandler(
+        "eth_signTransaction",
+        new InternalResponseHandler<>(
+            responseFactory,
+            new EthSignTransactionResultProvider(chainId, signerProvider, jsonDecoder)));
     return requestMapper;
   }